requirements.txt
python-dotenv~=1.2.1
mysql~=0.0.3
interactions.py~=5.16.0